requirements.txt
fastapi
uvicorn
pydantic
requests
typer
rich
httpx
python-dotenv
